agri_advisory_system/requirements.txt
fastapi
uvicorn
python-multipart
twilio
python-dotenv
requests
schedule
jinja2
